Add two new methods to generate PR information
Prior to this commit, the only functionality we had was generating
a hash of pull requests organized by their number and containing
their lifetime in minutes as well as whether or not they were merged.

Now, there are two new methods. One which simply generates an
array of pull request lifetimes (pull_request_lifetimes) and one
that calculates some useful data given a set of pull request
lifetimes (calculate averages).