Add PullRequest class
Prior to this commit, there was no class to represent pull requests,
and they were represented in a hash. Due to the need to store pull
request information into a database, a specific PullRequest class
was required. Now, that class has been added along with a method
to convert the pull request hash into an array of PullRequest
objects. Tests were also added.