Make local database connections work
Without this patch we cannot run a local Postgres database and test the
app quickly.  This is a problem because it's really slow to push to
Heroku each time we want to try a change.

Once this patch is applied the database connection to a local RDBMS
works, however we did run into one remaining issue.  The local postgres
server did not have the 'postgres' role defined for some reason.  We
resolved this by executing:

    createuser -h localhost -s -r postgres

This command creates the postgres role which ActiveRecord was
complaining about.  Once the role is in place the following command
starts a local instance of the web app:

    RACK_ENV=development foreman start