Stop writing repeated pull requests to database
Prior to this commit, even if a pull request was already in the
database, it would be written again. The reason this was happening
was because save_if_new was using the new_record? method. The reason
this method didn't work was because every time a new PullRequest object
was created, it would get a new id value automatically, making
new_record? return true when in fact the record was already in
the database.

Now the exists? method is used instead, because it allows you to
specify which fields to compare to determine if something is unique.