Remove undeeded code and update table
Prior to this commit, there was a 'lifetimes' column of the pull
request table. Now, it has been changed to have a "time opened"
and "time closed" columns since this will be more versatile when it
comes to organizing data.

Additionally, and method which generates an array of integers
representing pull request lifetimes has been removed as it is no
longer needed or relevant.