Update time_opened and time_closed columns and add graph
Use "datetime" instead of "date" for the time_closed and
time_opened columns in order to get a more accurate picture
of how long a pull request was open for.

Prior to this commit, there was no D3.js code in the project. Now
as an attempt to learn D3, a basic bar graph is generated using
SVG elements. Currently the graph is not using real data