Add check to avoid bad pull requests
Prior to this commit, there were issues with certain pull requests
because the user that opened them had deleted their account. This meant
that the needed data could not be retrieved and written to the database.

Now, there is a check before the call to write to the database is made
to ensure the needed fields are not null. If they are, it skips entry
of that pull request and prints a warning message.