Fix how "from_community" column is populated
Prior to this commit, the "from_community" column was not providing
accurate data since it was getting the "company" field from a user.

Now it checks if the user is a member of the Puppet Labs organization,
which means that even if the user has not accurately filled out their
"company" field, they will show up as a member of Puppet Labs.