Remove elastic y-axis option
Prior to this commit, those graphs that had a y-axis had an elastic
one meaning that it would change as the scale of the data changed.
This wasn't optimal because it made it harder to clearly see the
differences in the data being compared.

Now each graph as a static y-axis that will not change with the data.