Add open/closed column to pull_requests table
Prior to this commit there was no column which showed if a pull
request was opening or closed, since only closed pull requests
were being collected.

Now, there is a column to store the state of a pull request so
that information about currently open pull requests can also be
displayed.