Update Rake task to only run on Sundays
Since the Rake task to populate the database will now be run
everyday by Heroku, create a version of the job that will only
run on Sundays to avoid using unneeded time on Heroku.

Additionally, this commit adds a few lines of code that populate
the "closed" column of the "pull_requests" table.