Reorganize file structure
Prior to this commit, both the public and views directories were
stored in the lib directory. This was because Sinatra looked for them
in a specific place relative to the server file.

Now they have been relocated and the path Sinatra will use to locate
them has been reconfigured.

Additionally, this commit adds stubbing to the repository spec that
prevents tests from failing due to exceeded API rate limits.