Remove unneeded code
Prior to this commit, there was a lot of code in the project that
was no longer being used. Since the project has undergone a lot of
evolution from the time it was created, things had slipped by that
are not needed for the operation of the dashboard.

Now the unneeded code has been removed.