Refactor JavaScript and begin to add info buttons
Prior to this commit, all of the JavaScript for generating dc.js
graphs was all in once place and poorly organized. This made it
difficult to understand, harder to make changes, and impossible to
reuse the functionality elsewhere.

Now, the JavaScript has been refactored into methods and moved into
a separate file to fix the mentioned issues.

Information buttons have also been added to the first two graphs
on the dashboard. The buttons provide a much better way to give
a more detailed description about the graphs than the footnotes.