Add text input field and slider
Prior to this commit, there was no way for the user to define
what date range they wanted the graphs to represent. Now a text
field has been added as a temporary solution to this problem. If users
enter the start date in the right format, it will adjust the graphs
accordingly.

Due to the desire not to have to deal with sanitizing user input,
the text field will soon be replaced with a slider. The slider is
present on the page but does not adjust anything yet.