Add (poorly formatted) footer
Prior to this commit, the dashboard did not have a footer. This was
not idea because there was no clear place to put the slider. Now a
footer has been added and the slider has been put into it.

Currently, the footer is very poorly formatted and looks kind of awful.